Narrative_Engine/requirements.txt
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
matplotlib>=3.5.0
Pillow>=9.0.0
requests>=2.28.0
pydub>=0.25.0
